for now לא לעשות merge עדיין

+++ WebAPI/Controllers/ContactsController.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using System.Net;
 using WebAPI.Models;
 using WebAPI.Services;
 
@@ -8,6 +10,7 @@ namespace WebAPI.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+    //[Authorize]
     public class ContactsController : ControllerBase
     {
         private IContactService contactService;
@@ -20,13 +23,22 @@ namespace WebAPI.Controllers
             chatService = new ChatService();
         }
 
-        private void setMessageService(string id)
+        /*private void setChatService(string id)
+        {
+            chatService = new ChatService(id);
+        }*/
+
+        private int setMessageService(string id)
         {
             Chat c = chatService.Get(id);
-            messageService = new MessageService(c);
+            if (c != null)
+            {
+                messageService = new MessageService(c);
+                return 0;
+            }
+            return 1;
         }
 
-        // GET: api/<ContactsController>
         /// <summary>
         /// Returns all contacts.
         /// </summary>
@@ -34,10 +46,10 @@ namespace WebAPI.Controllers
         [HttpGet]
         public List<Contact> Get()
         {
+            Response.StatusCode = 200;
             return contactService.GetAllContacts();
         }
 
-        // GET api/<ContactsController>/:id
         /// <summary>
         /// Returns the contact with given id.
         /// </summary>
@@ -46,10 +58,17 @@ namespace WebAPI.Controllers
         [HttpGet("{id}")]
         public Contact Get(string id)
         {
-            return contactService.Get(id);
+            Contact contact = contactService.Get(id);
+            if (contact == null)
+            {
+                // Response.StatusCode = 404;
+                Response.StatusCode = (int) HttpStatusCode.NotFound;
+                return null;
+            }
+            Response.StatusCode = 200;
+            return contact;
         }
 
-        // GET api/<ContactsController>/:id/messages
         /// <summary>
         /// Returns all messages.
         /// </summary>
@@ -58,11 +77,22 @@ namespace WebAPI.Controllers
         [HttpGet("{id}/messages")]
         public List<Message> GetMessages(string id)
         {
-            setMessageService(id);
-            return messageService.GetAllMessages();
+            if (setMessageService(id) > 0)
+            {
+                Response.StatusCode = 404;
+                return null;
+            }
+            List<Message> messages = messageService.GetAllMessages();
+
+            if (messages == null)
+            {
+                Response.StatusCode = 404;
+                return null;
+            }
+            Response.StatusCode = 200;
+            return messages;
         }
 
-        // GET api/<ContactsController>/:id/messages/:id
         /// <summary>
         /// Returns a certain message sent in a chat with a certain contact.
         /// </summary>
@@ -72,11 +102,22 @@ namespace WebAPI.Controllers
         [HttpGet("{id1}/messages/{id2}")]
         public Message Get(string id1, string id2)
         {
-            setMessageService(id1);
-            return messageService.Get(id2);
+            if (setMessageService(id1) > 0)
+            {
+                Response.StatusCode = 404;
+                return null;
+            }
+            Message message = messageService.Get(id2);
+
+            if (message == null)
+            {
+                Response.StatusCode = 404;
+                return null;
+            }
+            Response.StatusCode = 200;
+            return message;
         }
 
-        // POST api/<ContactsController>
         /// <summary>
         /// Creates new contact.
         /// </summary>
@@ -84,10 +125,16 @@ namespace WebAPI.Controllers
         [HttpPost]
         public void Post([FromBody] RequestCreationOfNewContact request)
         {
-            chatService.CreateChat(request.Id, request.Name, request.Server);
+            // setChatService(Global.Id);
+            int num = chatService.CreateChat(request.Id, request.Name, request.Server);
+            if (num > 0)
+            {
+                Response.StatusCode = 404;
+                return;
+            }
+            Response.StatusCode = 201;
         }
 
-        // POST api/<ContactsController>/:id/messages
         /// <summary>
         /// Creates new message.
         /// </summary>
@@ -95,11 +142,15 @@ namespace WebAPI.Controllers
         [HttpPost("{id}/messages")]
         public void Post(string id, [FromBody] RequestCreationOfNewMessage request)
         {
-            setMessageService(id);
+            if (setMessageService(id) > 0)
+            {
+                Response.StatusCode = 404;
+                return;
+            }
             messageService.SendMessage(request.Content, true);
+            Response.StatusCode = 201;
         }
 
-        // PUT api/<ContactsController>/:id
         /// <summary>
         /// Updates a certain contact, by id.
         /// </summary>
@@ -108,12 +159,20 @@ namespace WebAPI.Controllers
         [HttpPut("{id}")]
         public void Put(string id, [FromBody] RequestEditContact request)
         {
-            contactService.Edit(id, request.Name, request.Server);
             Contact contact = contactService.Get(id);
+            if (contact == null)
+            {
+                Response.StatusCode = 404;
+                return;
+            }
+
+            contactService.Edit(id, request.Name, request.Server);
+            
+            // setChatService(Global.Id);
             chatService.Edit(id, contact);
+            Response.StatusCode = 204;
         }
 
-        // POST api/<ContactsController>/:id/messages/:id
         /// <summary>
         /// Updates a certain message sent in a chat with a certain contact.
         /// </summary>
@@ -123,12 +182,18 @@ namespace WebAPI.Controllers
         [HttpPut("{id1}/messages/{id2}")]
         public void Put(string id1, string id2, [FromBody] RequestEditMessage request)
         {
-            setMessageService(id1);
+            bool b1 = setMessageService(id1) > 0;
+            bool b2 = messageService.Get(id2) == null;
+            if (b1 || b2)
+            {
+                Response.StatusCode = 404;
+                return;
+            }
             messageService.Edit(id2, true, request.Content, DateTime.Now);
-            
+            Response.StatusCode = 204;
+
         }
 
-        // DELETE api/<ContactsController>/:id
         /// <summary>
         /// Deletes a certain contact, by id.
         /// </summary>
@@ -136,10 +201,15 @@ namespace WebAPI.Controllers
         [HttpDelete("{id}")]
         public void Delete(string id)
         {
+            if (chatService.Get(id) == null)
+            {
+                Response.StatusCode = 404;
+                return;
+            }
             chatService.Delete(id);
+            Response.StatusCode = 204;
         }
 
-        // DELETE api/<ContactsController>/:id/messages/:id
         /// <summary>
         /// Deletes a certain message sent in a chat with a certain contact.
         /// </summary>
@@ -148,19 +218,20 @@ namespace WebAPI.Controllers
         [HttpDelete("{id1}/messages/{id2}")]
         public void Delete(string id1, string id2)
         {
-            setMessageService(id1);
-            /*Chat c = chatService.Get(id1);
-            messageService = new MessageService(c);*/
+            bool b1 = setMessageService(id1) > 0;
+            bool b2 = messageService.Get(id2) == null;
+            if (b1 || b2)
+            {
+                Response.StatusCode = 404;
+                return;
+            }
+            
             messageService.Delete(id2);
 
             List<Message> messages = messageService.GetAllMessages();
-            if (messages.Count > 0)
-            {
-                /*Contact contact = contactService.Get(id1);*/
-                /*Contact contact = c.Contact;*/
-                contactService.UpdateLastDate(id1, messages);
+            contactService.UpdateLastDate(id1, messages);
 
-            }
+            Response.StatusCode = 204;
         }
     }
 }
